feat: refactor translation handling to move saving logic to translationService with caching
